Refactor helpers into separate module
