+++ tests/core/pyspec/eth2spec/test/lightclient_patch/sanity/test_blocks.py
@@ -1,6 +1,4 @@
 import random
-from eth2spec.test.helpers.keys import privkeys
-from eth2spec.utils import bls
 from eth2spec.test.helpers.state import (
     state_transition_and_sign_block,
     next_epoch,
@@ -8,39 +6,14 @@ from eth2spec.test.helpers.state import (
 from eth2spec.test.helpers.block import (
     build_empty_block_for_next_slot,
 )
+from eth2spec.test.lightclient_patch.helpers import (
+    compute_aggregate_sync_committee_signature,
+)
 from eth2spec.test.context import (
     PHASE0, PHASE1,
     with_all_phases_except,
     spec_state_test,
 )
-
-
-def compute_sync_committee_signature(spec, state, slot, privkey):
-    domain = spec.get_domain(state, spec.DOMAIN_SYNC_COMMITTEE, spec.compute_epoch_at_slot(slot))
-    if slot == state.slot:
-        block_root = build_empty_block_for_next_slot(spec, state).parent_root
-    else:
-        block_root = spec.get_block_root_at_slot(state, slot)
-    signing_root = spec.compute_signing_root(block_root, domain)
-    return bls.Sign(privkey, signing_root)
-
-
-def compute_aggregate_sync_committee_signature(spec, state, slot, participants):
-    if len(participants) == 0:
-        return spec.G2_POINT_AT_INFINITY
-
-    signatures = []
-    for validator_index in participants:
-        privkey = privkeys[validator_index]
-        signatures.append(
-            compute_sync_committee_signature(
-                spec,
-                state,
-                slot,
-                privkey,
-            )
-        )
-    return bls.Aggregate(signatures)
 
 
 def run_sync_committee_sanity_test(spec, state, fraction_full=1.0):
